+++ templates/demo-store/app/lib/type.ts
@@ -10,3 +10,12 @@ export type Locale = {
 };
 
 export type Localizations = Record<string, Locale>;
+
+export enum CartAction {
+  ADD_TO_CART = 'ADD_TO_CART',
+  REMOVE_FROM_CART = 'REMOVE_FROM_CART',
+  UPDATE_CART = 'UPDATE_CART',
+  UPDATE_DISCOUNT = 'UPDATE_DISCOUNT',
+  UPDATE_BUYER_IDENTITY = 'UPDATE_BUYER_IDENTITY',
+}
+export type CartActions = keyof typeof CartAction;
