+++ templates/demo-store/app/routes/cart.tsx
@@ -2,31 +2,18 @@ import {CartLoading, Cart} from '~/components';
 import {Await, useMatches} from '@remix-run/react';
 import {Suspense} from 'react';
 import invariant from 'tiny-invariant';
-import {json, type ActionArgs} from '@remix-run/oxygen';
+import {json, type ActionArgs, type AppLoadContext} from '@remix-run/oxygen';
 import type {
-  CartLineInput,
-  CartBuyerIdentityInput,
   Cart as CartType,
-  UserError,
-  CartUserError,
+  CartInput,
+  CartLineInput,
   CartLineUpdateInput,
+  CartUserError,
+  UserError,
+  CartBuyerIdentityInput,
 } from '@shopify/hydrogen-react/storefront-api-types';
-import {
-  cartCreate,
-  cartAdd,
-  cartDiscountCodesUpdate,
-  cartRemove,
-  cartUpdate,
-  cartUpdateBuyerIdentity,
-} from '~/data';
 import {isLocalPath} from '~/lib/utils';
-
-type CartAction =
-  | 'ADD_TO_CART'
-  | 'REMOVE_FROM_CART'
-  | 'UPDATE_CART'
-  | 'UPDATE_DISCOUNT'
-  | 'UPDATE_BUYER_IDENTITY';
+import {CartAction, type CartActions} from '~/lib/type';
 
 export async function action({request, context}: ActionArgs) {
   const {session, storefront} = context;
@@ -38,7 +25,7 @@ export async function action({request, context}: ActionArgs) {
     session.get('customerAccessToken'),
   ]);
 
-  const cartAction = formData.get('cartAction') as CartAction;
+  const cartAction = formData.get('cartAction') as CartActions;
   invariant(cartAction, 'No cartAction defined');
 
   const countryCode = formData.get('countryCode')
@@ -52,7 +39,7 @@ export async function action({request, context}: ActionArgs) {
   };
 
   switch (cartAction) {
-    case 'ADD_TO_CART':
+    case CartAction.ADD_TO_CART:
       const lines = formData.get('lines')
         ? (JSON.parse(String(formData.get('lines'))) as CartLineInput[])
         : ([] as CartLineInput[]);
@@ -77,7 +64,7 @@ export async function action({request, context}: ActionArgs) {
         });
       }
       break;
-    case 'REMOVE_FROM_CART':
+    case CartAction.REMOVE_FROM_CART:
       const lineIds = formData.get('linesIds')
         ? (JSON.parse(String(formData.get('linesIds'))) as CartType['id'][])
         : ([] as CartType['id'][]);
@@ -90,7 +77,7 @@ export async function action({request, context}: ActionArgs) {
       });
 
       break;
-    case 'UPDATE_CART':
+    case CartAction.UPDATE_CART:
       const updateLines = formData.get('lines')
         ? (JSON.parse(String(formData.get('lines'))) as CartLineUpdateInput[])
         : ([] as CartLineUpdateInput[]);
@@ -103,7 +90,7 @@ export async function action({request, context}: ActionArgs) {
       });
 
       break;
-    case 'UPDATE_DISCOUNT':
+    case CartAction.UPDATE_DISCOUNT:
       invariant(cartId, 'Missing cartId');
 
       const formDiscountCode = formData.get('discountCode');
@@ -115,7 +102,7 @@ export async function action({request, context}: ActionArgs) {
         storefront,
       });
       break;
-    case 'UPDATE_BUYER_IDENTITY':
+    case CartAction.UPDATE_BUYER_IDENTITY:
       const buyerIdentity = formData.get('buyerIdentity')
         ? (JSON.parse(
             String(formData.get('buyerIdentity')),
@@ -173,4 +160,335 @@ export default function CartRoute() {
       </Suspense>
     </div>
   );
+}
+
+/*
+  Cart Queries
+*/
+
+const USER_ERROR_FRAGMENT = `#graphql
+  fragment ErrorFragment on CartUserError {
+    message
+    field
+    code
+  }
+`;
+
+const LINES_CART_FRAGMENT = `#graphql
+  fragment CartLinesFragment on Cart {
+    id
+    totalQuantity
+  }
+`;
+
+//! @see: https://shopify.dev/api/storefront/2022-01/mutations/cartcreate
+const CREATE_CART_MUTATION = `#graphql
+  mutation ($input: CartInput!, $country: CountryCode = ZZ, $language: LanguageCode)
+  @inContext(country: $country, language: $language) {
+    cartCreate(input: $input) {
+      cart {
+        ...CartLinesFragment
+      }
+      errors: userErrors {
+        ...ErrorFragment
+      }
+    }
+  }
+  ${LINES_CART_FRAGMENT}
+  ${USER_ERROR_FRAGMENT}
+`;
+
+/**
+ * Create a cart with line(s) mutation
+ * @param input CartInput https://shopify.dev/api/storefront/2022-01/input-objects/CartInput
+ * @see https://shopify.dev/api/storefront/2022-01/mutations/cartcreate
+ * @returns result {cart, errors}
+ * @preserve
+ */
+export async function cartCreate({
+  input,
+  storefront,
+}: {
+  input: CartInput;
+  storefront: AppLoadContext['storefront'];
+}) {
+  const {cartCreate} = await storefront.mutate<{
+    cartCreate: {
+      cart: CartType;
+      errors: CartUserError[];
+    };
+    errors: UserError[];
+  }>(CREATE_CART_MUTATION, {
+    variables: {input},
+  });
+
+  invariant(cartCreate, 'No data returned from cartCreate mutation');
+
+  return cartCreate;
+}
+
+const ADD_LINES_MUTATION = `#graphql
+  mutation ($cartId: ID!, $lines: [CartLineInput!]!, $country: CountryCode = ZZ, $language: LanguageCode)
+  @inContext(country: $country, language: $language) {
+    cartLinesAdd(cartId: $cartId, lines: $lines) {
+      cart {
+        ...CartLinesFragment
+      }
+      errors: userErrors {
+        ...ErrorFragment
+      }
+    }
+  }
+  ${LINES_CART_FRAGMENT}
+  ${USER_ERROR_FRAGMENT}
+`;
+
+/**
+ * Storefront API cartLinesAdd mutation
+ * @param cartId
+ * @param lines [CartLineInput!]! https://shopify.dev/api/storefront/2022-01/input-objects/CartLineInput
+ * @see https://shopify.dev/api/storefront/2022-01/mutations/cartLinesAdd
+ * @returns result {cart, errors}
+ * @preserve
+ */
+export async function cartAdd({
+  cartId,
+  lines,
+  storefront,
+}: {
+  cartId: string;
+  lines: CartLineInput[];
+  storefront: AppLoadContext['storefront'];
+}) {
+  const {cartLinesAdd} = await storefront.mutate<{
+    cartLinesAdd: {
+      cart: CartType;
+      errors: CartUserError[];
+    };
+  }>(ADD_LINES_MUTATION, {
+    variables: {cartId, lines},
+  });
+
+  invariant(cartLinesAdd, 'No data returned from cartLinesAdd mutation');
+
+  return cartLinesAdd;
+}
+
+const REMOVE_LINE_ITEMS_MUTATION = `#graphql
+  mutation ($cartId: ID!, $lineIds: [ID!]!, $language: LanguageCode, $country: CountryCode)
+  @inContext(country: $country, language: $language) {
+    cartLinesRemove(cartId: $cartId, lineIds: $lineIds) {
+      cart {
+        id
+        totalQuantity
+        lines(first: 100) {
+          edges {
+            node {
+              id
+              quantity
+              merchandise {
+                ...on ProductVariant {
+                  id
+                }
+              }
+            }
+          }
+        }
+      }
+      errors: userErrors {
+        message
+        field
+        code
+      }
+    }
+  }
+`;
+
+/**
+ * Create a cart with line(s) mutation
+ * @param cartId the current cart id
+ * @param lineIds [ID!]! an array of cart line ids to remove
+ * @see https://shopify.dev/api/storefront/2022-07/mutations/cartlinesremove
+ * @returns mutated cart
+ * @preserve
+ */
+export async function cartRemove({
+  cartId,
+  lineIds,
+  storefront,
+}: {
+  cartId: string;
+  lineIds: CartType['id'][];
+  storefront: AppLoadContext['storefront'];
+}) {
+  const {cartLinesRemove} = await storefront.mutate<{
+    cartLinesRemove: {cart: CartType; errors: UserError[]};
+  }>(REMOVE_LINE_ITEMS_MUTATION, {
+    variables: {
+      cartId,
+      lineIds,
+    },
+  });
+
+  invariant(cartLinesRemove, 'No data returned from remove lines mutation');
+  return cartLinesRemove;
+}
+
+const LINES_UPDATE_MUTATION = `#graphql
+  ${LINES_CART_FRAGMENT}
+  ${USER_ERROR_FRAGMENT}
+  mutation ($cartId: ID!, $lines: [CartLineUpdateInput!]!, $language: LanguageCode, $country: CountryCode)
+  @inContext(country: $country, language: $language) {
+    cartLinesUpdate(cartId: $cartId, lines: $lines) {
+      cart {
+        ...CartLinesFragment
+      }
+      errors: userErrors {
+        ...ErrorFragment
+      }
+    }
+  }
+`;
+
+/**
+ * Update cart line(s) mutation
+ * @param cartId the current cart id
+ * @param lineIds [ID!]! an array of cart line ids to remove
+ * @see https://shopify.dev/api/storefront/2022-07/mutations/cartlinesremove
+ * @returns mutated cart
+ * @preserve
+ */
+export async function cartUpdate({
+  cartId,
+  lines,
+  storefront,
+}: {
+  cartId: string;
+  lines: CartLineUpdateInput[];
+  storefront: AppLoadContext['storefront'];
+}) {
+  const {cartLinesUpdate} = await storefront.mutate<{
+    cartLinesUpdate: {cart: CartType; errors: UserError[]};
+  }>(LINES_UPDATE_MUTATION, {
+    variables: {cartId, lines},
+  });
+
+  invariant(
+    cartLinesUpdate,
+    'No data returned from update lines items mutation',
+  );
+  return cartLinesUpdate;
+}
+
+/**
+ * @see https://shopify.dev/api/storefront/2022-10/mutations/cartBuyerIdentityUpdate
+ * @preserve
+ */
+const UPDATE_CART_BUYER_COUNTRY = `#graphql
+ mutation(
+   $cartId: ID!
+   $buyerIdentity: CartBuyerIdentityInput!
+   $country: CountryCode = ZZ
+   $language: LanguageCode
+ ) @inContext(country: $country, language: $language) {
+   cartBuyerIdentityUpdate(cartId: $cartId, buyerIdentity: $buyerIdentity) {
+     cart {
+       id
+       buyerIdentity {
+         email
+         phone
+         countryCode
+       }
+     }
+     errors: userErrors {
+       message
+       field
+       code
+     }
+   }
+ }
+`;
+
+/**
+ * Mutation to update a cart buyerIdentity
+ * @param cartId  Cart['id']
+ * @param buyerIdentity CartBuyerIdentityInput
+ * @returns {cart: Cart; errors: UserError[]}
+ * @see API https://shopify.dev/api/storefront/2022-10/mutations/cartBuyerIdentityUpdate
+ * @preserve
+ */
+export async function cartUpdateBuyerIdentity({
+  cartId,
+  buyerIdentity,
+  storefront,
+}: {
+  cartId: string;
+  buyerIdentity: CartBuyerIdentityInput;
+  storefront: AppLoadContext['storefront'];
+}) {
+  const {cartBuyerIdentityUpdate} = await storefront.mutate<{
+    cartBuyerIdentityUpdate: {cart: CartType; errors: UserError[]};
+  }>(UPDATE_CART_BUYER_COUNTRY, {
+    variables: {
+      cartId,
+      buyerIdentity,
+    },
+  });
+
+  invariant(
+    cartBuyerIdentityUpdate,
+    'No data returned from cart buyer identity update mutation',
+  );
+
+  return cartBuyerIdentityUpdate;
+}
+
+const DISCOUNT_CODES_UPDATE = `#graphql
+  mutation cartDiscountCodesUpdate($cartId: ID!, $discountCodes: [String!], $country: CountryCode = ZZ)
+    @inContext(country: $country) {
+    cartDiscountCodesUpdate(cartId: $cartId, discountCodes: $discountCodes) {
+      cart {
+        id
+        discountCodes {
+          code
+        }
+      }
+      errors: userErrors {
+        field
+        message
+      }
+    }
+  }
+`;
+
+/**
+ * Mutation that updates the cart discounts
+ * @param discountCodes Array of discount codes
+ * @returns mutated cart
+ * @preserve
+ */
+export async function cartDiscountCodesUpdate({
+  cartId,
+  discountCodes,
+  storefront,
+}: {
+  cartId: string;
+  discountCodes: string[];
+  storefront: AppLoadContext['storefront'];
+}) {
+  const {cartDiscountCodesUpdate} = await storefront.mutate<{
+    cartDiscountCodesUpdate: {cart: CartType; errors: UserError[]};
+  }>(DISCOUNT_CODES_UPDATE, {
+    variables: {
+      cartId,
+      discountCodes,
+    },
+  });
+
+  invariant(
+    cartDiscountCodesUpdate,
+    'No data returned from the cartDiscountCodesUpdate mutation',
+  );
+
+  return cartDiscountCodesUpdate;
 }
